Fix clippy lints

+++ src/puzzle/scrambler.rs
@@ -93,7 +93,7 @@ where
     Piece: PrimInt,
 {
     fn scramble_with_rng<R: Rng>(&self, puzzle: &mut P, rng: &mut R) {
-        let n = puzzle.num_pieces() as usize;
+        let n = puzzle.num_pieces();
         let cycle_len = (self.length as usize).min(if n % 2 == 0 { n - 1 } else { n });
         let max = if cycle_len % 2 == 0 { n - 2 } else { n };
         let pieces = rand::seq::index::sample(rng, max, cycle_len);
